Updated tsis4 directory
Changed math's i.py a little bit (changed type from integer to float, because degrees can be written with with decimal points). Added all tasks from json.md file.

+++ tsis4/math/i.py
@@ -1,4 +1,4 @@
 import math
 
 a = input("Input degree: ")
-print("Output radian:", round(math.radians(int(a)), 6))
+print("Output radian:", round(math.radians(float(a)), 6))
